+++ objc2/src/macros.rs
@@ -1,3 +1,4 @@
+mod __rewrite_self_arg;
 mod declare_class;
 mod extern_class;