+++ objc2/src/macros/extern_class.rs
@@ -7,15 +7,21 @@
 /// `NSAutoreleasePool` does not have this!)
 ///
 /// You must specify the superclass of this class, similar to how you would
-/// in Objective-C. Due to Rust trait limitations, specifying e.g. the
-/// superclass `NSData` would not give you easy access to `NSObject`'s
-/// functionality, therefore you may specify additional parts of the
-/// inheritance chain.
+/// in Objective-C.
+///
+/// Due to Rust trait limitations, specifying e.g. the superclass `NSData`
+/// would not give you easy access to `NSObject`'s functionality. Therefore,
+/// you may specify additional parts of the inheritance chain using the
+/// `#[inherits(...)]` attribute.
 ///
 /// [`Encoding::Object`]: crate::Encoding::Object
 ///
 ///
 /// # Specification
+///
+/// The syntax is similar enough to Rust syntax that if you invoke the macro
+/// with parentheses (as opposed to curly brackets), [`rustfmt` will be able to
+/// format the contents][rustfmt-macros].
 ///
 /// This creates an opaque struct containing the superclass (which means that
 /// auto traits are inherited from the superclass), and implements the
@@ -23,6 +29,7 @@
 ///
 /// - [`RefEncode`][crate::RefEncode]
 /// - [`Message`][crate::Message]
+/// - [`ClassType`][crate::ClassType]
 /// - [`Deref<Target = $superclass>`][core::ops::Deref]
 /// - [`DerefMut`][core::ops::DerefMut]
 /// - [`AsRef<$inheritance_chain>`][AsRef]
@@ -30,13 +37,10 @@
 /// - [`Borrow<$inheritance_chain>`][core::borrow::Borrow]
 /// - [`BorrowMut<$inheritance_chain>`][core::borrow::BorrowMut]
 ///
-/// An associated function `class` is created on the object as a convenient
-/// shorthand so that you can do `MyObject::class()` instead of
-/// `class!(MyObject)`.
-///
 /// The macro allows specifying fields on the struct, but _only_ zero-sized
 /// types like [`PhantomData`] and [`declare::Ivar`] are allowed here!
 ///
+/// [rustfmt-macros]: https://github.com/rust-lang/rustfmt/discussions/5437
 /// [`PhantomData`]: core::marker::PhantomData
 /// [`declare::Ivar`]: crate::declare::Ivar
 ///
@@ -50,7 +54,7 @@
 /// [`NSObject`]: crate::foundation::NSObject
 ///
 ///
-/// # Example
+/// # Examples
 ///
 /// Create a new type to represent the `NSFormatter` class.
 ///
@@ -62,15 +66,19 @@
 /// # #[cfg(feature = "gnustep-1-7")]
 /// # unsafe { objc2::__gnustep_hack::get_class_to_force_linkage() };
 ///
-/// extern_class! {
+/// extern_class!(
 ///     /// An example description.
-///     #[derive(PartialEq, Eq, Hash)] // Uses `NSObject`'s implementation
-///     // Specify class and superclass
-///     // In this case the class `NSFormatter`, which subclasses `NSObject`
-///     unsafe pub struct NSFormatter: NSObject;
-/// }
+///     #[derive(PartialEq, Eq, Hash)] // Uses the superclass' implementation
+///     // Specify the class and struct name to be used
+///     pub struct NSFormatter;
 ///
-/// // Provided by the macro (it implements `ClassType`)
+///     // Specify the superclass, in this case `NSObject`
+///     unsafe impl ClassType for NSFormatter {
+///         type Superclass = NSObject;
+///     }
+/// );
+///
+/// // Provided by the implementation of `ClassType`
 /// let cls = NSFormatter::class();
 ///
 /// // `NSFormatter` implements `Message`:
@@ -81,20 +89,28 @@
 /// declared previously to specify as its superclass.
 ///
 /// ```
-/// use objc2::extern_class;
 /// use objc2::foundation::NSObject;
+/// use objc2::{extern_class, ClassType};
 /// #
-/// # extern_class! {
+/// # extern_class!(
 /// #     #[derive(PartialEq, Eq, Hash)]
-/// #     unsafe pub struct NSFormatter: NSObject;
-/// # }
+/// #     pub struct NSFormatter;
+/// #
+/// #     unsafe impl ClassType for NSFormatter {
+/// #         type Superclass = NSObject;
+/// #     }
+/// # );
 ///
-/// extern_class! {
+/// extern_class!(
 ///     #[derive(PartialEq, Eq, Hash)]
-///     // Specify the correct inheritance chain
-///     // `NSDateFormatter` subclasses `NSFormatter` which subclasses `NSObject`
-///     unsafe pub struct NSDateFormatter: NSFormatter, NSObject;
-/// }
+///     pub struct NSDateFormatter;
+///
+///     unsafe impl ClassType for NSDateFormatter {
+///         // Specify the correct inheritance chain
+///         #[inherits(NSObject)]
+///         type Superclass = NSFormatter;
+///     }
+/// );
 /// ```
 ///
 /// See the source code of `objc2::foundation` in general for more examples.
@@ -103,26 +119,46 @@
 macro_rules! extern_class {
     (
         $(#[$m:meta])*
-        unsafe $v:vis struct $name:ident: $superclass:ty $(, $inheritance_rest:ty)*;
+        $v:vis struct $name:ident;
+
+        unsafe impl ClassType for $for:ty {
+            $(#[inherits($($inheritance_rest:ty)+)])?
+            type Superclass = $superclass:ty;
+        }
     ) => {
         // Just shorthand syntax for the following
-        $crate::extern_class! {
+        $crate::extern_class!(
             $(#[$m])*
-            unsafe $v struct $name: $superclass $(, $inheritance_rest)* {}
-        }
+            $v struct $name {}
+
+            unsafe impl ClassType for $for {
+                $(#[inherits($($inheritance_rest)+)])?
+                type Superclass = $superclass;
+            }
+        );
     };
     (
         $(#[$m:meta])*
-        unsafe $v:vis struct $name:ident: $superclass:ty $(, $inheritance_rest:ty)* {
+        $v:vis struct $name:ident {
             $($field_vis:vis $field:ident: $field_ty:ty,)*
         }
+
+        unsafe impl ClassType for $for:ty {
+            $(#[inherits($($inheritance_rest:ty)+)])?
+            type Superclass = $superclass:ty;
+        }
     ) => {
-        $crate::__inner_extern_class! {
+        $crate::__inner_extern_class!(
             $(#[$m])*
-            unsafe $v struct $name<>: $superclass $(, $inheritance_rest)* {
+            $v struct $name<> {
                 $($field_vis $field: $field_ty,)*
             }
-        }
+
+            unsafe impl<> ClassType for $for {
+                $(#[inherits($($inheritance_rest)+)])?
+                type Superclass = $superclass;
+            }
+        );
 
         const _: () = {
             if $crate::__macro_helpers::size_of::<$name>() != 0 {
@@ -139,9 +175,13 @@ macro_rules! extern_class {
 #[doc(hidden)]
 #[macro_export]
 macro_rules! __impl_as_ref_borrow {
-    ($name:ident<$($t:ident $(: $b:ident)?),*>,) => {};
-    ($name:ident<$($t:ident $(: $b:ident)?),*>, $item:ty, $($tail:ty,)*) => {
-        impl<$($t $(: $b)?),*> $crate::__macro_helpers::AsRef<$item> for $name<$($t),*> {
+    {
+        impl ($($t:tt)*) for $for:ty;
+    } => {};
+    {
+        impl ($($t:tt)*) for $for:ty; $item:ty, $($tail:ty,)*
+    } => {
+        impl<$($t)*> $crate::__macro_helpers::AsRef<$item> for $for {
             #[inline]
             fn as_ref(&self) -> &$item {
                 // Triggers Deref coercion depending on return type
@@ -149,7 +189,7 @@ macro_rules! __impl_as_ref_borrow {
             }
         }
 
-        impl<$($t $(: $b)?),*> $crate::__macro_helpers::AsMut<$item> for $name<$($t),*> {
+        impl<$($t)*> $crate::__macro_helpers::AsMut<$item> for $for {
             #[inline]
             fn as_mut(&mut self) -> &mut $item {
                 // Triggers DerefMut coercion depending on return type
@@ -163,7 +203,7 @@ macro_rules! __impl_as_ref_borrow {
         // In particular, `Eq`, `Ord` and `Hash` all give the same results
         // after borrow.
 
-        impl<$($t $(: $b)?),*> $crate::__macro_helpers::Borrow<$item> for $name<$($t),*> {
+        impl<$($t)*> $crate::__macro_helpers::Borrow<$item> for $for {
             #[inline]
             fn borrow(&self) -> &$item {
                 // Triggers Deref coercion depending on return type
@@ -171,7 +211,7 @@ macro_rules! __impl_as_ref_borrow {
             }
         }
 
-        impl<$($t $(: $b)?),*> $crate::__macro_helpers::BorrowMut<$item> for $name<$($t),*> {
+        impl<$($t)*> $crate::__macro_helpers::BorrowMut<$item> for $for {
             #[inline]
             fn borrow_mut(&mut self) -> &mut $item {
                 // Triggers Deref coercion depending on return type
@@ -179,7 +219,9 @@ macro_rules! __impl_as_ref_borrow {
             }
         }
 
-        $crate::__impl_as_ref_borrow!($name<$($t $(: $b)?),*>, $($tail,)*);
+        $crate::__impl_as_ref_borrow! {
+            impl ($($t)*) for $for; $($tail,)*
+        }
     };
 }
 
@@ -189,19 +231,28 @@ macro_rules! __inner_extern_class {
     // TODO: Expose this variant in the `object` macro.
     (
         $(#[$m:meta])*
-        unsafe $v:vis struct $name:ident<$($t:ident $(: $b:ident $(= $default:ty)?)?),*>: $superclass:ty $(, $inheritance_rest:ty)* {
+        $v:vis struct $name:ident<$($t_struct:ident $(: $b_struct:ident $(= $default:ty)?)?),*> {
             $($field_vis:vis $field:ident: $field_ty:ty,)*
+        }
+
+        unsafe impl<$($t_for:ident $(: $b_for:ident)?),*> ClassType for $for:ty {
+            $(#[inherits($($inheritance_rest:ty)+)])?
+            type Superclass = $superclass:ty;
         }
     ) => {
         $crate::__inner_extern_class! {
             @__inner
             $(#[$m])*
-            unsafe $v struct $name<$($t $(: $b $(= $default)?)?),*>: $superclass, $($inheritance_rest,)* $crate::runtime::Object {
+            $v struct $name ($($t_struct $(: $b_struct $(= $default)?)?),*) {
                 $($field_vis $field: $field_ty,)*
+            }
+
+            unsafe impl ($($t_for $(: $b_for)?),*) for $for {
+                INHERITS = [$superclass, $($($inheritance_rest,)+)? $crate::runtime::Object];
             }
         }
 
-        unsafe impl<$($t $(: $b)?),*> $crate::ClassType for $name<$($t),*> {
+        unsafe impl<$($t_for $(: $b_for)?),*> ClassType for $for {
             type Superclass = $superclass;
 
             #[inline]
@@ -212,15 +263,20 @@ macro_rules! __inner_extern_class {
     };
     (
         @__inner
+
         $(#[$m:meta])*
-        unsafe $v:vis struct $name:ident<$($t:ident $(: $b:ident $(= $default:ty)?)?),*>: $superclass:ty $(, $inheritance_rest:ty)* {
+        $v:vis struct $name:ident ($($t_struct:tt)*) {
             $($field_vis:vis $field:ident: $field_ty:ty,)*
+        }
+
+        unsafe impl ($($t:tt)*) for $for:ty {
+            INHERITS = [$superclass:ty $(, $inheritance_rest:ty)*];
         }
     ) => {
         $(#[$m])*
         // TODO: repr(transparent) when the inner pointer is no longer a ZST.
         #[repr(C)]
-        $v struct $name<$($t $(: $b $(= $default)?)?),*> {
+        $v struct $name<$($t_struct)*> {
             __inner: $superclass,
             // Additional fields (should only be zero-sized PhantomData or ivars).
             $($field_vis $field: $field_ty,)*
@@ -232,7 +288,7 @@ macro_rules! __inner_extern_class {
         //   that it actually inherits said object.
         // - The rest of the struct's fields are ZSTs, so they don't influence
         //   the layout.
-        unsafe impl<$($t $(: $b)?),*> $crate::RefEncode for $name<$($t),*> {
+        unsafe impl<$($t)*> $crate::RefEncode for $for {
             const ENCODING_REF: $crate::Encoding<'static>
                 = <$superclass as $crate::RefEncode>::ENCODING_REF;
         }
@@ -243,7 +299,7 @@ macro_rules! __inner_extern_class {
         //
         // That the object must work with standard memory management is upheld
         // by the caller.
-        unsafe impl<$($t $(: $b)?),*> $crate::Message for $name<$($t),*> {}
+        unsafe impl<$($t)*> $crate::Message for $for {}
 
         // SAFETY: An instance can always be _used_ in exactly the same way as
         // its superclasses (though not necessarily _constructed_ in the same
@@ -263,7 +319,7 @@ macro_rules! __inner_extern_class {
         // Note that you can easily have two different variables pointing to
         // the same object, `x: &T` and `y: &T::Target`, and this would be
         // perfectly safe!
-        impl<$($t $(: $b)?),*> $crate::__macro_helpers::Deref for $name<$($t),*> {
+        impl<$($t)*> $crate::__macro_helpers::Deref for $for {
             type Target = $superclass;
 
             #[inline]
@@ -283,28 +339,30 @@ macro_rules! __inner_extern_class {
         // But `&mut NSMutableString` -> `&mut NSString` safe, since the
         // `NSCopying` implementation of `NSMutableString` is used, and that
         // is guaranteed to return a different object.
-        impl<$($t $(: $b)?),*> $crate::__macro_helpers::DerefMut for $name<$($t),*> {
+        impl<$($t)*> $crate::__macro_helpers::DerefMut for $for {
             #[inline]
             fn deref_mut(&mut self) -> &mut Self::Target {
                 &mut self.__inner
             }
         }
 
-        impl<$($t $(: $b)?),*> $crate::__macro_helpers::AsRef<Self> for $name<$($t),*> {
+        impl<$($t)*> $crate::__macro_helpers::AsRef<Self> for $for {
             #[inline]
             fn as_ref(&self) -> &Self {
                 self
             }
         }
 
-        impl<$($t $(: $b)?),*> $crate::__macro_helpers::AsMut<Self> for $name<$($t),*> {
+        impl<$($t)*> $crate::__macro_helpers::AsMut<Self> for $for {
             #[inline]
             fn as_mut(&mut self) -> &mut Self {
                 self
             }
         }
 
-        $crate::__impl_as_ref_borrow!($name<$($t $(: $b)?),*>, $superclass, $($inheritance_rest,)*);
+        $crate::__impl_as_ref_borrow! {
+            impl ($($t)*) for $for; $superclass, $($inheritance_rest,)*
+        }
     };
 }
 
@@ -355,30 +413,48 @@ macro_rules! __attribute_helper {
         $($fn)*
     };
 
-    // Extract and convert the `#[sel(...)]` attribute to a `sel!` invocation.
+    // Extract the `#[sel(...)]` attribute and send it to another macro
     {
         @extract_sel
-        #[sel($($sel:tt)*)]
-        $($rest:tt)*
+        ($out_macro:path)
+        (
+            #[sel($($sel:tt)*)]
+            $($rest:tt)*
+        )
+        $($macro_args:tt)*
     } => {{
         $crate::__attribute_helper! {
             @extract_sel_duplicate
             $($rest)*
         }
 
-        $crate::sel!($($sel)*)
+        $out_macro!(
+            $($macro_args)*
+            @($($sel)*)
+        )
     }};
     {
         @extract_sel
-        #[$($m_checked:tt)*]
-        $($rest:tt)*
+        ($out_macro:path)
+        (
+            #[$($m_checked:tt)*]
+            $($rest:tt)*
+        )
+        $($macro_args:tt)*
     } => {{
         $crate::__attribute_helper! {
             @extract_sel
-            $($rest)*
+            ($out_macro)
+            ($($rest)*)
+            $($macro_args)*
         }
     }};
-    {@extract_sel} => {{
+    {
+        @extract_sel
+        ($out_macro:path)
+        ()
+        $($macro_args:tt)*
+    } => {{
         compile_error!("Must specify the desired selector using `#[sel(...)]`");
     }};