+++ objc2/src/rc/test_object.rs
@@ -48,13 +48,17 @@ std::thread_local! {
     pub(crate) static TEST_DATA: RefCell<ThreadTestData> = RefCell::new(Default::default());
 }
 
-declare_class! {
+declare_class!(
     /// A helper object that counts how many times various reference-counting
     /// primitives are called.
     #[derive(Debug, PartialEq)]
-    unsafe pub(crate) struct RcTestObject: NSObject {}
+    pub(crate) struct RcTestObject {}
 
-    unsafe impl {
+    unsafe impl ClassType for RcTestObject {
+        type Superclass = NSObject;
+    }
+
+    unsafe impl RcTestObject {
         #[sel(alloc)]
         fn alloc() -> *mut Self {
             TEST_DATA.with(|data| data.borrow_mut().alloc += 1);
@@ -123,7 +127,7 @@ declare_class! {
             Id::consume_as_ptr(ManuallyDrop::new(Self::new()))
         }
     }
-}
+);
 
 unsafe impl Send for RcTestObject {}
 unsafe impl Sync for RcTestObject {}